update namespace of Minetools::Facetool::Face class. (for creating new classes.)
