+++ app/controllers/api/v1/texture/face_controller.rb
@@ -5,7 +5,7 @@ class Api::V1::Texture::FaceController < ApplicationController
 			return nil
 		end
 
-		@face = Mineface::Face.new name: params[:id]
+		@face = Minetools::Facetool::Face.new name: params[:id]
 		begin
 			@face.request!
 			@image_bin = @face.image.to_blob
@@ -19,7 +19,7 @@ class Api::V1::Texture::FaceController < ApplicationController
 	end
 
 	private def steve_face_image
-		@face = Mineface::Face.new
+		@face = Minetools::Facetool::Face.new
 		return @face.get_face_image(Rails.root.join('public', "steve.png").to_s)
 	end
 end
